Move the creation of ViewHolder for the header and footer view out of HeaderFooterAdapter
The reason is that RecyclerView implementation may invoke onCreateViewHolder for header/footer
multiple times even though there is only one header/footer view in the list. When it happens,
per the implementation before this change, the same header/footer view will be bound to
multiple ViewHolders, and this will mass up the references between the View and ViewHolder
inside RecyclerView, causing IllegalArgumentException("Called attach on a child which is not
detached").

After this change, onCreateViewHolder will return a new ViewHolder which always contains a
new View.